Replace template PPP overview
The default template PPP (teams/johndoe.md) contains references to AddressBook-3.
We keep this default template as a reference for team members' own PPP. At the
same time, the overview section is modified to refer to HRMate instead, so that
team members can copy this template without having to write their own overviews.