Add spacing between bullet points
MarkBind unfortunately wraps main bullet points in a paragraph
tag. This has the effect of leaving a disproportionate amount
of padding between main bullet points and nested bullet points.

To fix this, I changed main.css so that all bullet points will
have a padding of 0.5rem, which makes the padding look more
proportionate.

The site nav menu also uses an unlisted list, hence it must be
specifically selected so that its padding is reset to 0rem.