Implement import and export commands.
The import command triggers a file dialog that allows the user to
select the CSV file to import. Users are likely to type in a file
path wrongly, hence the choice of a file dialog.

The export command allows the user to input a file name to save the
records to. The export command parser ensures that file names
provided are valid and will save to the "/exports" folder in CSV
format.

Currently, the import command only triggers a file dialog and
displays the name of the file selected, while the export command
creates an empty file with the file name provided. Future work
includes implementing the CSV parser to write the address book
in CSV format, as well as parse the CSV file into the address
book.