Refactor CsvFile::addRow() method to accept CsvParsable interface
Previously, the method required the user to supply a List, from which
string values can be obtained by invoking the list's objects' toString()
method. However, there may be some objects for which the toString()
method does not return the desired string representation.

By changing the parameter to be of CsvParsable type, objects that
want to serialize into CSV must implement CsvParsable::getCsvValues(),
which returns a List of Strings. This provides greater control over
the string representation of the values.