Integrate Leaves Storage system with Main application
* Tweak UserPrefs to include directory for reading Leaves json file
* Tweak MainApp to initialise LeavesBook
* Tweak ModelManager and StorageManager to receive LeavesBook in
constructor.

Notes:

* Based on Storage interface, we can now interact with the Leaves json file
enabling read and write. However, I think that in order for it to work properly,
one must first delete the preferences.json file, so that MainApp will recreate the
new preferences.json with the required field specifying the LeavesBook datapath
* Do also note that I had intentionally left in the original constructors for
StorageManager and ModelManager for now. Work is not complete, and I do not wish to
break any existing functionality or tests. I will remove them once done.
* My work extends to the storage function and only that for now. Therefore, I have not
extended much functionalities on the ModelManager and Model interface side.