(WIP) Implement unit tests for JsonSerializableLeavesBook and JsonAdaptedLeave classes
Current situation:
* Prior to converting a JsonAdaptedLeave object into a Leave object to be stored
in the LeavesBook, there must be a check at some point in time during the
conversion with the AddressBook to ensure that the person whose name is noted
in the Leave object corresponds to a record in the address book itself.

* Going into this implementation, I plan to leave this check to be done at as
high an abstraction level as possible, as opposed to passing the addressbook into
the JsonAdaptedLeave to check things. However, without passing the addressbook into
JsonAdaptedLeave, I am unable to convert the name into a corresponding Person object
in the JsonAdaptedLeave level.

* Instead, I have opted to expose these fields to the JsonSerializableLeavesBook level
instead, which will take the addressBook in the toModelType()
method to convert the JsonAdaptedLeave objects into Leave objects

* By doing so, I have broken the layers of abstraction in the JsonAdaptedLeave class,
and as a result I am unable to properly test it. Do note that the current toModelType()
method of the JsonAdaptedLeave class also does not work and returns null.

* This current block prevents me from writing the remaining unit tests for these two classes