Add delimiter specifier in generated CSV files
The semicolon is an uncommon separator, and other spreadsheet applications
may have difficulty identifying it and instead look for a comma to delimit.

By adding sep=; in the first line of the CSV file, other applications can
recognise the separator and parse the file correctly, thus requiring no
further configurations in order to read the file.

Additionally, by identifying the delimiter from the first line of the CSV
file, non-semicolon-delimited files can also be read into HRMate.