Implement import and export of leaves
ExportParsers and Export and Import commands were reused as much
as possible.

Changed behaviour: If the person in a leave record does not correspond
to any person in the address book, the entry is skipped instead of
throwing an Exception - this allows the other valid leaves in the
LeavesBook file to be imported.