Standardise error messages for index
With the change in error handling of ParserUtil, error messages
for commands with indices must be updated to reflect the correct
nature of the error. Overflow and out-of-bound indices have their
own error messages.

Additional changes include modifying the message contents to improve
helpfulness, and modifying Messages::format(Leave) to only show
the description if there is a non-empty description.