Fix leaves' employees not updating their names
Leave::belongsTo previously used equals, which is an issue as there
is no guarantee that the employee field in the leave contains the
same instance as the employee list. By changing it to use
ComparablePerson::isSamePerson instead, only the name will be
compared, thus fixing the bug.

Test cases were added for Leave and EditCommandIntegrationTest
to guarantee that the code produces the desired behaviour.